pyproject.toml
[project]
name = "rag-knowledgebase"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "beautifulsoup4>=4.14.3",
    "ddgs>=9.12.1",
    "duckduckgo-search>=8.1.1",
    "faiss-cpu>=1.13.2",
    "fastapi>=0.135.3",
    "huggingface-hub>=1.8.0",
    "langchain>=1.2.14",
    "langchain-community>=0.4.1",
    "langchain-core>=1.2.24",
    "langchain-experimental>=0.4.1",
    "langchain-groq>=1.1.2",
    "langchain-huggingface>=1.2.1",
    "langchain-neo4j>=0.9.0",
    "langchain-ollama>=1.0.1",
    "langgraph>=1.1.4",
    "neo4j>=6.1.0",
    "pymupdf>=1.27.2.2",
    "pypdf>=6.9.2",
    "python-dotenv>=1.2.2",
    "python-multipart>=0.0.22",
    "sentence-transformers>=5.3.0",
    "transformers>=5.4.0",
    "uvicorn[standard]>=0.42.0",
    "wikipedia>=1.4.0",
]
